refactor(payment): separate external confirm call and DB transaction, split tests (#186)
